requirements.txt
python-dotenv
groq
pymupdf
streamlit
pillow